PCP-24 Implement reconnection logic
Here we build on the heartbeat sending work to reconnect to the broker when
either the initial connection fails, or the connection is determined to have
been closed.

The major behavioural difference here is the move to asyncronous
connections, so though the return value of `(puppetlabs.pcp.client/connect)`
is still a Client, it may not be a connected client.  In order to
determine if the client is connected you may wait for it to become
connected with the newly added `(puppetlabs.pcp.client/wait-for-connected client timeout)`
helper, or poll the `(puppetlabs.pcp.client/open?  client)` predicate.

`(puppetlabs.pcp.client/send! client message)` may now throw a
`:puppetlabs.pcp.client/not-connected` exception, if the client is currently
connected.

Internal renamings:
  :conn -> :websocket-connection
  :websocket -> :websocket-client
  :heartbeat-stop -> :should-stop