PCP-24 avoid race condition in on-connect sending
In some configurations we see a race condition between the on-connect
handler firing and calling (send! client associaton-message) and
the future that returns the connection completing and updating the
client's :websocket-connection future atom.

Here we bypass this race by using the session passed to the on-connect
handler to directly send the association message to the remote end
of the websocket session.