PCP-43 enable use of with-open
Here we refactor a little to add a protocol that a pcp client will
implement (ClientInterface) and make Client be implemented as a
record type that implements this protocol.

This allows us to make use of with-open to ensure the close function is called
when we're done with a client, and more cleanly delinates the border between
client API and client implementation.

Also we've reordered some of the functions to match what seemed to be a more
logical narrative for the order of the ClientInterface, and ensured that all
functions are now schematised.