PCP-4 Handle session association in the client
Here we add tracking in for session association in the style of the
existing connection tracking.  This adds `wait-for-assocation`,
`associated?`, and `associate-response-received?` which behave
similarly to `wait-for-connection` and `open?`.

As a side effect of this a user can no longer specify their own
`http://puppetlabs.com/associate_response` handler, as we force our
internal one to be used over a user-supplied one, so we can be sure
that our callback is the one that gets used.