PCP-59 remove :state atom, determine state by inspection
Here we rework the way we track state, eliminating some duplicate state
tracking that was causing race conditions.

Now we can reliably see if a client is connected by checking the state of the
future in the :websocket-connection atom, and the state of association by
checking the promise in the :associate-response atom.

We remove the additional states of :closing and :closed, rename open? to
connected? and so now just support `connecting?` `connected?` `associating?`
`associated?` to inspect client state.