PCP-323 perform hostname verifcation of the broker
Here we change the configuration of the SslContextFactory to use "HTTPS"
endpoint verification, that verifies the hostname of the websockets uri against
the peer certificate from the server.

In order to allow this to be tested, we regenerate the test fixure
ssl certificates, and switch to using the 'localhost' certificate
in the captive broker.

SSL verification failure is treated in the same way as connection
refused.