(PCP-488) Disable schema validations by default
This commit removes the `^:always-validate` attribute from all functions
in this projects which used it in order to disable the schema validation
for those functions by default.
To compensate for that the tests in this project now use the
`schema.test/validate-schemas` fixture so that they still run with the
validations enabled.