(maint) Fix race condition in tests
The state of a dummy PCP client used by some tests in this project
was racy as it contained a `(future )` field which may have or
may have not been realized at the time the tests were testing it,
which was causing occasional test failures.
This commit makes sure the state of the dummy PCP client is stable
(i.e. its `(future )` field is realized) before its handed over to
the tests.